requirements.txt
googlemaps
google-generativeai
python-dotenv
rich
